internal code restructuring
-move the text messages to a separate file
-move argparser to a separate module
-rename the 'keyPath' cmd param to the 'tlsKeyPath'
